requirements.txt
Django==1.6
South==0.8.1
amqp==1.0.13
ansi2html==0.9.3
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.32
celery==3.0.23
django-celery==3.0.23
django-crispy-forms==1.4.0
django-redis==3.3
django-tastypie==0.11.0
docker_py==0.3.2
kombu==2.5.14
mimeparse==0.1.3
mock==1.0.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2013b
redis==2.7.6
requests==1.2.3
six==1.5.2
times==0.6.1
websocket-client==0.11.0
wsgiref==0.1.2

fabric==1.8.1
paramiko==1.12.1
PyYAML==3.10